requirements.txt
fastapi
uvicorn
boto3
numpy
tensorflow
scikit-learn
pillow
prometheus_client